hashTable in cpp

+++ hash_in_cpp/src/prime.cpp
@@ -1,5 +1,5 @@
-#include <limits.h>
-#include <math.h>
+#include <limits>
+#include <cmath>
 
 #include "prime.h"
 
@@ -16,7 +16,7 @@ size_t is_prime(const size_t x){
   if(x<2) return -1;
   if(x<4) return 1;
   if((x%2) == 0) return 0;
-  for(int i=3;i <= floor(sqrt((double)x)); i+=2){
+  for(int i=3;i <= floor(std::sqrt(x)); i+=2){
     if(x%i==0)
       return 0;
   }
